requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
scikit-learn
pyyaml
pandas
tqdm
